+++ post-install.d/flatpak.sh
@@ -1,7 +1,10 @@
 # bash script to be sourced from fedora_setup.sh
 
 # Add flathub repository and delete fedora's repos
-if sudo flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo >/dev/null; then
+if
+sudo flatpak --system remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo >/dev/null && \
+	sudo flatpak --user remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo >/dev/null && \
+then
 	sudo flatpak remote-delete fedora >/dev/null
 	sudo flatpak remote-delete fedora-testing >/dev/null
 fi